+++ testing/matrix-sdk-test/src/sync_builder/mod.rs
@@ -252,13 +252,21 @@ impl SyncResponseBuilder {
 
 /// Helper trait to mutate the data in [`State`].
 trait StateMutExt {
+    /// Use the `After` variant rather than `Before`.
+    fn use_state_after(&mut self);
+    /// Access the inner list of state events.
     fn events_mut(&mut self) -> &mut Vec<Raw<AnySyncStateEvent>>;
 }
 
 impl StateMutExt for State {
+    fn use_state_after(&mut self) {
+        *self = Self::After(Default::default());
+    }
+
     fn events_mut(&mut self) -> &mut Vec<Raw<AnySyncStateEvent>> {
         match self {
             Self::Before(state) => &mut state.events,
+            Self::After(state) => &mut state.events,
             // We don't allow to construct another variant.
             _ => unreachable!(),
         }